Improve hover behavior when not enough space
This changes the hover's position fallback logic from:

If not enough room on right
  Position left
If not enough room on left
  Position right

To:

If not enough room on right
  If enough room on left
    Position left
  Else
    Position below
If not enough room on left
   If enough room on right
     Position right
   Else
     Position below

A new option showHoverHint is also added as part of this which will show a hint
to the user to hold alt/option in order to mouse over the hover (only when the
evaluated hover is true). This is an explicit setting such that it doesn't show
up on hovers such as the activity bar which looks awkward.

Fixes #176704